rename weather_json to weather
